fix: set up proxy earlier
this:
- makes code simpler
- also adds proxy for Neutron RPC
